Use i32 as Score

+++ crates/types/src/matcher.rs
@@ -1,5 +1,5 @@
 /// Score of base matching algorithm(fzy, skim, etc).
-pub type Score = i64;
+pub type Score = i32;
 
 /// A tuple of (score, matched_indices) for the line has a match given the query string.
 #[derive(Debug, Clone, PartialEq, Eq)]
